pull from upstream

+++ j2k/testData/fileOrElement/issues/kt-851.kt
@@ -1,9 +1,8 @@
 internal class Test {
-    fun putInt(i: Int) {
-    }
+    fun putInt(i: Int) {}
 
     fun test() {
         val b = 10
-        putInt(b.toInt())
+        putInt(b)
     }
 }
